+++ src/test/scala/eventstore/StreamSubscriptionActorSpec.scala
@@ -99,7 +99,7 @@ class StreamSubscriptionActorSpec extends AbstractSubscriptionActorSpec {
     "catch events that appear in between reading and subscribing" in new SubscriptionScope {
       connection expectMsg readEvents(0)
 
-      val position = 1
+      val position = 1L
       actor ! readCompleted(2, false, event0, event1)
 
       expectEvent(event0)
@@ -154,7 +154,7 @@ class StreamSubscriptionActorSpec extends AbstractSubscriptionActorSpec {
     "stop catching events that appear in between reading and subscribing if stop received" in new SubscriptionScope {
       connection expectMsg readEvents(0)
 
-      val position = 1
+      val position = 1L
       actor ! readCompleted(2, false, event0, event1)
 
       expectEvent(event0)
@@ -179,7 +179,7 @@ class StreamSubscriptionActorSpec extends AbstractSubscriptionActorSpec {
     }
 
     "continue with subscription if no events appear in between reading and subscribing" in new SubscriptionScope {
-      val position = 0
+      val position = 0L
       connection expectMsg readEvents(position)
       actor ! readCompleted(position, endOfStream = true)
 
@@ -198,7 +198,7 @@ class StreamSubscriptionActorSpec extends AbstractSubscriptionActorSpec {
 
     "continue with subscription if no events appear in between reading and subscribing and position is given" in
       new SubscriptionScope {
-        val position = 1
+        val position = 1L
         connection expectMsg readEvents(position)
 
         actor ! readCompleted(position, endOfStream = true)
@@ -216,7 +216,7 @@ class StreamSubscriptionActorSpec extends AbstractSubscriptionActorSpec {
       }
 
     "forward events while subscribed" in new SubscriptionScope {
-      val position = 0
+      val position = 0L
       connection expectMsg readEvents(position)
       actor ! readCompleted(position, endOfStream = true)
 
@@ -424,12 +424,12 @@ class StreamSubscriptionActorSpec extends AbstractSubscriptionActorSpec {
 
     def expectEvent(x: Event) = expectMsg(x)
 
-    def newEvent(number: Int): Event = EventRecord(streamId, EventNumber.Exact(number), mock[EventData])
+    def newEvent(number: Long): Event = EventRecord(streamId, EventNumber.Exact(number), mock[EventData])
 
-    def readEvents(x: Int) =
+    def readEvents(x: Long) =
       ReadStreamEvents(streamId, EventNumber(x), readBatchSize, Forward, resolveLinkTos = resolveLinkTos)
 
-    def readCompleted(next: Int, endOfStream: Boolean, events: Event*) = ReadStreamEventsCompleted(
+    def readCompleted(next: Long, endOfStream: Boolean, events: Event*) = ReadStreamEventsCompleted(
       events = events.toList,
       nextEventNumber = EventNumber(next),
       lastEventNumber = mock[EventNumber.Exact],
@@ -438,6 +438,6 @@ class StreamSubscriptionActorSpec extends AbstractSubscriptionActorSpec {
       direction = Forward
     )
 
-    def subscribeCompleted(x: Int) = SubscribeToStreamCompleted(x.toLong, Some(EventNumber.Exact(x)))
+    def subscribeCompleted(x: Long) = SubscribeToStreamCompleted(x.toLong, Some(EventNumber.Exact(x)))
   }
 }
